DEV: Move chat selection buttons out of chat-live-pane
We are adding another button here and potentially more
in future, it makes more sense to have the chat selection
management buttons to be in their own component, not clogging
up chat-live-pane.
